Update to starter pack extension
This commit includes changes required to upgrade
from the unsupported starter pack version we are
using to the supported, automatically upgradable
starter pack extension.

The 0.4.0 version is pinned so that we can
validate the changes coming in and avoid any
breaking changes while were are in the middle of
other documentation changes.